backend/requirements.txt
Django>=5.0,<6.0
requests
django-cors-headers
djangorestframework